backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=6.0
django-cors-headers>=4.0
